// src/app/models/car-detail.ts
export class CarDetail {
    id?: string;
    name?: string;
    model?: string;
    year?: string;
    gear?: string;
    seat?: string;
    color?: string;
    owner?: string;
    fuel?: string;
    milage?: string;
    cc?: string;
    address?: string;
    user?: string;
    usage?: string;
    carEntry?: string;
    brand?: string;
    type?: string;
    price?: string;
    imageUrl?: string;
    video?: string;
}